Change to rect - 2 element rects, and argument name.
We can now create a rect from a two tuple, which gives x,y and w,h
that tuple. This is useful for cases where SDL uses just the size
or just the position from a rect.

to_sdl_rect can now take an argument name, which we can use to make
it clear which argument is wrong.