Use pygame's blit code.
Pygame has a weird default blit mode, where it ignores the src alpha
channel when blending onto a pixel with a destination alpha of 0.
While not the usual blend mode, this provides better results by
default in a lot of cases.

This patch also enables all of the other blend modes, including the
PREMULTIPLY mode that makes the most sense.

Unfortunately, it means that pygame_sdl2 now has LGPL-ed code in it,
so that means pygame_sdl2 is completely SDL2-licensed.