Remove use of SDL2_gfx in favor of included code.
This isn't really SDL2_gfx's fault. It uses SDL2's Renderers, which
seem to have a number of problems when used to draw to a surface that
doesn't start fully opaque. (It looks like many of the functions do
not update the destination alpha.)

This replaces the use of Renderers with the same SDL_gfx code that was
embedded into pygame, so now everything should work the same as
before.

Fixes renpy/renpy#612.