Match blit behavior to pygame.
This means that we used the pygame alpha blit function of the
surface has source alpha, or a special blend mode is used, and
we use the SDL blit function otherwise.

This allows us to get color key and alpha modulation correct
without any sort of special handling.

Fixes #23 and improves the fix to #20.